+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 dev
 ---
 
+* Changed the default value of the `ignore_nicks` option from `''` to
+  `'-,--,-->'`. This will prevent notifications for messages from nicks `-`,
+  `--`, and `-->`, which is usually just noise
+  ([#23](https://github.com/s3rvac/weechat-notify-send/issues/23)).
 * Added a new option: `notify_on_all_messages_in_current_buffer`. When set to
   `on`, you will receive a notification on all messages in the currently
   active buffer